requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
watchfiles==1.0.0
pydantic==2.9.1
playwright==1.47.0
pillow==10.4.0
requests==2.32.3
numpy==1.26.4
google-generativeai==0.8.1
anthropic==0.39.0
